refactor: SearchOption 타입 통합 및 SearchDialog 필터에서 불필요한 타입 제거
