+++ src/types.ts
@@ -1,10 +1,10 @@
 export interface Lecture {
-  id: string;
+  id: string; // 과목코드
   title: string;
   credits: string;
-  major: string;
-  schedule: string;
-  grade: number;
+  major: string; // 전공
+  schedule: string; // 시간
+  grade: number; // 학년
 }
 
 export interface Schedule {
@@ -12,4 +12,13 @@ export interface Schedule {
   day: string;
   range: number[];
   room?: string;
+}
+
+export interface SearchOption {
+  query?: string;
+  grades: number[];
+  days: string[];
+  times: number[];
+  majors: string[];
+  credits?: number;
 }
